SCC line break missing (#217)
